# api/openapi.yaml
openapi: 3.1.0
info:
  title: Better Call AI API
  version: 0.1.0
  description: 1인 1사건 소유권과 사용자 확인을 중심으로 한 MVP 계약
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/cases/{caseId}/files/{fileId}/legal-evidence/search:
    parameters:
      - $ref: '#/components/parameters/CaseId'
      - name: fileId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: searchCaseEvidence
      summary: 선택한 파일의 현재 확정 OCR 발췌로 법률 근거 후보 검색
      description: >
        내 사건의 현재 확정 수정본 1개에서 최대 650 UTF-16 코드 단위를 발췌해 질문과 함께 검색한다.
        기본 발췌 시작은 0이며 전체 문서 검색이 아니다. 응답에 사용한 수정본 ID와 실제 발췌 범위를 표시한다.
        질문과 발췌문은 용어 보완 후 OpenAI 임베딩으로 전송된다. 원본 파일·미확정 초안·다른 파일은 전송하지 않는다.
        외부 호출 전후 소유권·사건 버전·현재 수정본을 확인한다. 도중 변경 시 결과를 폐기하지만 이미 수행한 호출 비용은 발생할 수 있다.
        결과는 마지막 확인 시점의 버전 기준이며 응답 이후 변경을 막는 잠금은 아니다. 법률 답변 생성이나 결과 저장은 하지 않는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, expectedCaseVersion]
              properties:
                query: { type: string, minLength: 1, maxLength: 300, description: 공백만 입력할 수 없음 }
                expectedCaseVersion: { type: integer, minimum: 1 }
                excerptStart:
                  type: [integer, 'null']
                  minimum: 0
                  description: 생략/null이면 0. 확정 수정본의 UTF-16 시작 위치이며 문자 중간이나 본문 끝 이후는 400
      responses:
        '200':
          description: 현재 확정본 발췌와 근거 후보
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success: { type: boolean, const: true }
                  traceId: { type: [string, 'null'] }
                  data:
                    type: object
                    required: [caseId, caseVersion, evidence, results]
                    properties:
                      caseId: { type: string, format: uuid }
                      caseVersion: { type: integer, minimum: 1 }
                      evidence: { $ref: '#/components/schemas/EvidenceExcerpt' }
                      results:
                        type: object
                        required: [items, page, pageSize, hasNext]
                        properties:
                          items:
                            type: array
                            maxItems: 8
                            items: { $ref: '#/components/schemas/LegalEvidence' }
                          page: { type: integer, const: 1 }
                          pageSize: { type: integer, const: 8 }
                          hasNext: { type: boolean, const: false }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: CASE_001 내 활성 사건이 아님 또는 FILE_001 해당 사건·파일에 사용 가능한 현재 확정 수정본이 없음
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409':
          description: CASE_002 사건 버전 불일치 또는 외부 호출 중 현재 확정본 변경·제거
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429':
          description: LEGAL_DATA_003 공통 검색 동시 처리 한도
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '503':
          description: COMMON_002 키 미설정 또는 LEGAL_DATA_002 검색 실패
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
  /api/v1/cases/{caseId}/chat/turns:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      operationId: sendChatTurn
      summary: 질문 선저장 및 완성 AI 답변 저장 후 반환
      description: >
        사건 정리 전용이며 RAG 법률 판단은 미지원이다. 질문을 먼저 커밋하고 외부 호출 동안 DB 연결을 반환한다.
        답변 저장 커밋 이후에만 200으로 응답한다. 동일 사건의 동일 키/질문 재전송은 재생성하지 않는다.
        키에 다른 질문을 사용하면 409, 활성 요청/서버 보호 제한은 429이다.
        실패 시 목록을 조회하고 retryAllowed일 때 명시적 재시도 API를 사용한다.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: 사건별 턴 ID로 사용할 클라이언트 UUID. 네트워크 재전송 시 같은 값 유지
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 4000, description: 공백만 입력할 수 없음 }
      responses:
        '200':
          description: 완성 답변 저장 완료 또는 기존 답변 재전송
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatTurnSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteFile
      summary: 내 사건의 자료 삭제
      description: 자료를 논리 삭제하고 현재 분석을 최신 아님으로 표시한다. 임시 원본은 삭제 작업으로 즉시 정리하며 실패하면 백그라운드 정리가 재시도한다.
      responses:
        '204':
          description: 자료 삭제 완료
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ChatConflict' }
        '429': { $ref: '#/components/responses/ChatBusy' }
        '503': { $ref: '#/components/responses/ChatUnavailable' }
    get:
      operationId: listChatTurns
      summary: 질문과 저장 완료된 답변의 최신순 페이지 조회
      description: >
        내부 실행 상태는 반환하지 않는다. answer가 null이면 질문만 표시한다.
        retryAllowed가 true인 마지막 질문에만 재시도 버튼을 표시한다.
        3분 지난 실행은 다음 접근에서 실패 처리하며 자동 재생성하지 않는다.
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 최신 턴부터 반환. 프론트는 화면 배치 시 역순 정렬
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatTurnPageSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ChatUnavailable' }
...
  /api/v1/auth/register:
    post:
      operationId: register
      summary: 로컬 계정 회원가입 및 Token Pair 발급
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: 계정 생성됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: 이미 가입된 이메일
  /api/v1/auth/login:
    post:
      operationId: login
      summary: 로컬 계정 로그인
      description: >
        정규화 이메일별 연속 실패 5회부터 15분간 429 AUTH_008을 반환한다.
        미가입 이메일에도 동일하게 적용하며 잠금 중 요청은 만료 시각을 연장하지 않는다.
        성공 또는 잠금 만료 후 실패 횟수를 초기화한다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429':
          description: 로그인 실패 제한 (AUTH_008). 잠금이 해제될 때까지 재시도를 중단한다.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh Token 1회성 회전
      description: >
        동일 토큰의 동시 재발급도 재사용으로 판단하여 해당 계정의 모든 Refresh Token을 폐기한다.
        클라이언트는 재발급 요청을 한 번만 실행하며 AUTH_006이면 다시 로그인한다.
        이미 발급된 Access Token은 만료 전까지 유효하다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: 새 Token Pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: 현재 Refresh Token 폐기
      description: >
        반복 요청은 204로 처리하며 다른 사용자의 토큰에는 영향을 주지 않는다.
        이미 발급된 Access Token은 만료 전까지 유효하다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '204':
          description: 로그아웃 처리됨
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/me:
    get:
      operationId: getMe
      summary: 인증 사용자 정보 조회
      responses:
        '200':
          description: 내 사용자 정보
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/cases:
    get:
      operationId: listCases
      summary: 내 활성 사건 목록 조회
      description: >
        삭제되지 않은 본인 사건만 updatedAt 내림차순, 동일 시각에는 id 내림차순으로 반환한다.
        긴 진술과 목표 본문은 제외한다. 전체 건수 대신 hasNext를 제공하며 페이지는 1부터 시작한다.
        페이지 사이 사건이 추가·수정·삭제되면 위치가 바뀔 수 있으므로 목록을 새로 조회한다.
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, maximum: 10000, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 사건 요약 목록. 유효한 페이지 값이지만 해당 위치에 사건이 없으면 빈 배열과 hasNext=false
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CasePageSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createCase
      summary: 사건 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/cases/{caseId}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    delete:
      operationId: deleteCase
      summary: 내 사건 논리 삭제
      description: >
        삭제 표시와 CASE_DELETED 이벤트를 하나의 트랜잭션에 기록한다.
        사건 및 연결 데이터는 즉시 접근 차단되며 기본 7일 뒤 영구 삭제 가능한 시각을 저장한다.
        영구 삭제 배치 및 객체 저장소 파기 작업은 아직 미구현이므로 204는 물리 삭제 완료를 의미하지 않는다.
        현재 버전을 기준으로 삭제하며 반복 삭제, 없는 사건, 타인 사건은 모두 404를 반환한다.
      responses:
        '204':
          description: 논리 삭제 완료. 응답 본문 없음
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      operationId: getCase
      summary: 내 사건 조회
      responses:
        '200':
          description: 사건
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseSuccess' }
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      operationId: updateCase
      summary: 사건 진술 수정 및 기존 분석 stale 처리
      description: >
        소유자는 JWT sub로만 식별한다. 입력 검증을 통과한 요청에서 타인 소유 또는 없는 사건은 기대 버전 일치 여부와 관계없이 404를 반환한다.
        소유자 요청에서만 오래된 버전을 409로 구분하며 거부된 요청은 사건과 outbox를 변경하지 않는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        '200':
          description: 변경됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseSuccess' }
        '409':
          description: version 충돌
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/cases/{caseId}/files:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      operationId: listCaseFiles
      summary: 내 사건의 파일 메타데이터 목록
      description: 제거된 파일 제외, 원본 삭제된 파일 포함. 생성 시각·ID 내림차순. 원본 경로와 OCR 본문은 반환하지 않는다.
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, maximum: 10000, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: 공통 페이지 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  traceId: { type: [string, 'null'] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/FileResponse' } }
                      page: { type: integer }
                      pageSize: { type: integer }
                      hasNext: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: CASE_001 내 활성 사건이 아님 }
    post:
      operationId: uploadFile
      summary: 내 사건의 원본 파일 임시 업로드
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
          description: 사용자별 요청 UUID. 동일 내용 재시도에 같은 키를 사용한다.
      description: >
        PDF·PNG·JPEG·WebP만 허용하며 확장자, MIME, 헤더와 파싱 결과를 확인한다.
        파일당 20MiB, PDF 30쪽, 사건당 10개·100MiB가 기본 한도다.
        ClamAV 검사 후 형식 파싱을 진행하며 신규 성공 응답의 malwareStatus는 clean이다.
        검사 실패 시 우회하지 않고 업로드 전체를 거부한다.
        원본 다운로드나 OCR 실행을 의미하지 않는다.
        무료는 UTC 하루 10회이며 유료에는 일일 횟수·총용량 상한이 없다.
        예약 후 실패도 무료 횟수에 포함된다. 동일 키 재요청은 재검사·추가 차감 없이
        기존 성공 파일의 현재 메타데이터 또는 저장된 실패 코드를 반환한다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: 임시 저장 완료 및 메타데이터 반환
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 사건 한도 초과 (FILE_005), 동일 키의 내용 불일치 (FILE_010), 동일 요청 처리 중 (FILE_011)
        '413':
          description: 파일 20MiB 또는 multipart 요청 21MiB 초과 (FILE_003)
        '422':
          description: 악성 또는 검사 정책상 안전하지 않은 파일 (FILE_007)
        '429':
          description: 무료 플랜 일일 업로드 횟수 초과 (FILE_009)
        '503':
          description: 임시 저장소 사용 불가 (FILE_006) 또는 보안 검사 장애·동시 처리 한도 초과 (FILE_008)
  /api/v1/cases/{caseId}/files/{fileId}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
      - name: fileId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getFile
      summary: 내 사건의 파일 메타데이터 조회
      description: 타인 파일, 다른 사건의 파일, 삭제된 사건의 파일은 모두 404. 원본 경로와 다운로드 URL은 반환하지 않는다.
      responses:
        '200':
          description: 파일 메타데이터
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CaseId:
      name: caseId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    ChatConflict:
      description: CHAT_001 키에 다른 질문 사용 또는 CHAT_006 재시도 세대/대상 충돌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ChatBusy:
      description: CHAT_002 동시 요청/처리 제한. 잠시 후 목록 조회. 새 UUID를 자동 발급해 재전송하지 않음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ChatUnavailable:
      description: COMMON_002 연동 미설정, CHAT_003 생성 실패, CHAT_005 저장 확인 실패. 목록 조회 후 명시적 재시도
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: JWT가 없거나 서명·만료·발급자·대상자 검증에 실패함
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 없거나 접근할 수 없는 자원
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    EvidenceExcerpt:
      type: object
      required: [fileId, revisionId, text, start, end, totalLength, partial]
      properties:
        fileId: { type: string, format: uuid }
        revisionId: { type: string, format: uuid }
        text: { type: string, maxLength: 650 }
        start: { type: integer, minimum: 0, description: 확정 수정본 UTF-16 시작 위치(포함) }
        end: { type: integer, minimum: 1, description: 확정 수정본 UTF-16 끝 위치(미포함) }
        totalLength: { type: integer, minimum: 1, description: 확정 수정본 전체 UTF-16 길이 }
        partial: { type: boolean, description: 앞이나 뒤에 사용하지 않은 본문이 있으면 true }
    LegalEvidence:
      type: object
      required: [chunkId, documentId, title, heading, content, sourceUrl, versionLabel, effectiveFrom, rankScore]
      properties:
        chunkId: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        title: { type: string }
        heading: { type: string }
        content: { type: string }
        sourceUrl: { type: string, format: uri }
        versionLabel: { type: string }
        effectiveFrom: { type: string, format: date }
        rankScore: { type: number, description: RRF 정렬 점수이며 확률·법률적 신뢰도가 아님 }
...
    ChatTurn:
      type: object
      required: [id, question, answer, attempt, retryAllowed, createdAt, answeredAt]
      properties:
        id: { type: string, format: uuid }
        question: { type: string, maxLength: 4000 }
        answer: { type: [string, 'null'], maxLength: 16000, description: 저장 완료된 답변만 제공. HTML로 직접 삽입 금지 }
        attempt: { type: integer, minimum: 1 }
        retryAllowed: { type: boolean, description: 마지막 질문의 명시적 재시도 허용 여부. 내부 실행 상태는 아님 }
        createdAt: { type: string, format: date-time }
        answeredAt: { type: [string, 'null'], format: date-time }
    ChatTurnSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/ChatTurn' }
        traceId: { type: [string, 'null'] }
    ChatTurnPageSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data:
          type: object
          required: [items, page, pageSize, hasNext]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ChatTurn' }
            page: { type: integer }
            pageSize: { type: integer }
            hasNext: { type: boolean }
        traceId: { type: [string, 'null'] }
    RegisterRequest:
      type: object
      required: [email, password, displayName, termsAccepted, privacyAccepted]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, minLength: 8, maxLength: 72 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        termsAccepted: { type: boolean, const: true }
        privacyAccepted: { type: boolean, const: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, maxLength: 72 }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 1, maxLength: 500 }
    AuthToken:
      type: object
      required: [tokenType, accessToken, accessTokenExpiresAt, refreshToken, refreshTokenExpiresAt]
      properties:
        tokenType: { type: string, const: Bearer }
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
        refreshToken: { type: string }
        refreshTokenExpiresAt: { type: string, format: date-time }
    User:
      type: object
      required: [id, email, displayName]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
    AuthTokenSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/AuthToken' }
        traceId: { type: [string, 'null'] }
    UserSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/User' }
        traceId: { type: [string, 'null'] }
    CreateCaseRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 200 }
        disputeDomain: { type: string, enum: [housing_lease, vehicle_accident, assault, labor, consumer, commercial, family, inheritance, defamation, personal_injury], default: housing_lease }
        userPartyRole: { type: [string, 'null'], maxLength: 40 }
        userGoal: { type: [string, 'null'], maxLength: 2000 }
        originalStatement: { type: [string, 'null'], maxLength: 50000 }
    UpdateCaseRequest:
      type: object
      required: [originalStatement, expectedVersion]
      properties:
        originalStatement: { type: string, minLength: 1, maxLength: 50000 }
        expectedVersion: { type: integer, minimum: 1 }
        userPartyRole: { type: [string, 'null'], maxLength: 40, description: 생략하면 기존 역할 유지 }
        userGoal: { type: [string, 'null'], maxLength: 2000, description: 생략하면 기존 목표 유지 }
    Case:
      type: object
      required: [id, title, disputeDomain, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        disputeDomain: { type: string, enum: [housing_lease, vehicle_accident, assault, labor, consumer, commercial, family, inheritance, defamation, personal_injury] }
        status: { type: string, enum: [DRAFT, COLLECTING, REVIEW_READY, ANALYZING, READY, CLOSED] }
        userPartyRole: { type: [string, 'null'] }
        userGoal: { type: [string, 'null'] }
        originalStatement: { type: [string, 'null'] }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CaseSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/Case' }
        traceId: { type: [string, 'null'] }
    CaseSummary:
      type: object
      required: [id, title, disputeDomain, status, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        disputeDomain: { type: string, enum: [housing_lease, vehicle_accident, assault, labor, consumer, commercial, family, inheritance, defamation, personal_injury] }
        status: { type: string, enum: [DRAFT, COLLECTING, REVIEW_READY, ANALYZING, READY, CLOSED] }
        userPartyRole: { type: [string, 'null'] }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CasePage:
      type: object
      required: [items, page, pageSize, hasNext]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CaseSummary' }
        page: { type: integer, minimum: 1, maximum: 10000 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        hasNext: { type: boolean }
    CasePageSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/CasePage' }
        traceId: { type: [string, 'null'] }
...
    FileResponse:
      type: object
      required: [id, caseId, originalName, mimeType, sizeBytes, lifecycleStatus, malwareStatus, purgeStatus, createdAt, storageExpiresAt]
      properties:
        id: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        originalName: { type: string, maxLength: 255 }
        mimeType: { type: string, enum: [application/pdf, image/png, image/jpeg, image/webp] }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 20971520 }
        pageCount: { type: [integer, 'null'], minimum: 1, maximum: 30 }
        lifecycleStatus: { type: string }
        malwareStatus: { type: string }
        purgeStatus: { type: string }
        createdAt: { type: string, format: date-time }
        storageExpiresAt: { type: string, format: date-time }
    FileSuccess:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/FileResponse' }
        traceId: { type: [string, 'null'] }
...
    ApiError:
      type: object
      required: [success, code, message, occurredAt]
      properties:
        success: { type: boolean, const: false }
        code: { type: string }
        message: { type: string }
        traceId: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }